fix: change data types to double and add lookup file export
- Changed stoichiometric coefficients, reaction bounds, and objective storage from `int` to `double` to support decimal values in SBML.
- Fixed significant model errors caused by incorrect integer assumptions.
- Added export of lookup file for exchange/demand reactions to facilitate medium file construction.

+++ src/SBML_objective.cpp
@@ -1,5 +1,6 @@
 #include "SBML_objective.h"
 #include <cstdlib>
+#include <string>
 
 bool SBML::objective::read(rapidxml::xml_node<>* node, reaction* react) {
 	rapidxml::xml_node<>* obj_node = node->first_node("fbc:objective");
@@ -16,7 +17,7 @@ bool SBML::objective::read(rapidxml::xml_node<>* node, reaction* react) {
 			auto coef_attr = flux_node->first_attribute("fbc:coefficient");
 			if (!coef_attr) { return false; }
 
-			m_objective.push_back({ itr->second, atoi(coef_attr->value()) });
+			m_objective.push_back({ itr->second, std::stod(coef_attr->value()) });
 
 			flux_node = flux_node->next_sibling();
 		}
